Integrated @marc-villanueva stuff with @fbatista PR
