+++ lib/tire.rb
@@ -28,7 +28,6 @@ require 'tire/search/sort'
 require 'tire/search/facet'
 require 'tire/search/filter'
 require 'tire/search/highlight'
-require 'tire/search/suggest'
 require 'tire/search/scan'
 require 'tire/search/script_field'
 require 'tire/suggest'
